Analytics service refactoring
