+++ backend/beyond_the_loop/models/companies.py
@@ -128,6 +128,15 @@ class CompanyResponse(BaseModel):
 
 
 class CompanyTable:
+    def get_all(self):
+        try:
+            with get_db() as db:
+                companies = db.query(Company).all()
+                return [CompanyModel.model_validate(company) for company in companies]
+        except Exception as e:
+            print(f"Error getting companies: {e}")
+            return None
+
     def get_company_by_id(self, company_id: str):
         try:
             with get_db() as db:
@@ -357,33 +366,5 @@ class CompanyTable:
         except Exception as e:
             print(f"Error calculating credit limit for company {company_id}: {e}")
             return 1  # Default fallback value
-
-def crm_sync_companies_adoption_rate():
-    try:
-        thirty_days_ago = int((datetime.now() - timedelta(days=30)).timestamp())
-        requests_per_second = 25
-        batch_size = requests_per_second
-
-        with get_db() as db:
-            companies = db.query(Company).all()
-            companies = [CompanyModel.model_validate(company) for company in companies]
-
-            for i in range(0, len(companies), batch_size):
-                batch_start_time = time.time()
-                batch = companies[i:i + batch_size]
-
-                for company in batch:
-                    total_users = len(get_users_by_company(company.id))
-                    active_users = len(get_active_users_by_company(company.id, thirty_days_ago))
-                    adoption_rate = round((active_users / total_users * 100), 2) if total_users > 0 else 0
-
-                    crm_service.update_company_adoption_rate(company_name=company.name, adoption_rate=adoption_rate)
-
-                batch_processing_time = time.time() - batch_start_time
-                if batch_processing_time < 1.0:
-                    time.sleep(1.0 - batch_processing_time)
-
-    except Exception as e:
-        print(f"Error calculating companies adoption rate: {e}")
 
 Companies = CompanyTable()
